+++ crates/devela/src/media/font/bitmap/_.rs
@@ -6,8 +6,9 @@ crate::mods_in! {
     #[cfg(test)]
     mod _test;
 
-    mod view; // FontBitmapView, GlyphBitmapView
-    mod word; // FontBitmapWord
+    mod pixel;
+    mod view;
+    mod word;
 
     mod fonts; // Fonts::BIT_3_3,  Fonts::BIT_3_5, Fonts::BIT_5_6
     mod termivela; // Fonts::TERMIVELA_*
@@ -15,8 +16,9 @@ crate::mods_in! {
 crate::mods_out! { // _mods
     _mods {
         pub use super::{
-            view::*,
-            word::*,
+            pixel::{FontBitmapPixel, FontBitmapPixelIter},
+            view::{FontBitmapView, GlyphBitmapView},
+            word::FontBitmapWord,
         };
     }
 }